requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.8.0
apispec==6.6.1
apispec-webframeworks==1.1.0
flask-swagger-ui==4.11.1
PyJWT==2.9.0
argon2-cffi==23.1.0
marshmallow==3.21.3
psycopg[binary]==3.2.13
python-dotenv==1.0.1
requests==2.32.3
redis==5.0.8
minio==7.2.7
gunicorn==22.0.0
pytest==8.3.2
